+++ packages/models/windows_cpu_time_total.json
@@ -1,1 +1,1 @@
-{"growth": "linear", "n_changepoints": 3, "specified_changepoints": false, "changepoint_range": 0.8, "yearly_seasonality": "auto", "weekly_seasonality": "auto", "daily_seasonality": "auto", "seasonality_mode": "multiplicative", "seasonality_prior_scale": 10.0, "changepoint_prior_scale": 0.05, "holidays_prior_scale": 10.0, "mcmc_samples": 0, "interval_width": 0.8, "uncertainty_samples": 1000, "y_scale": 13034.611328125, "logistic_floor": false, "country_holidays": null, "component_modes": {"additive": ["additive_terms", "extra_regressors_additive"], "multiplicative": ["multiplicative_terms", "extra_regressors_multiplicative", "holidays"]}, "changepoints": "{\"name\":\"ds\",\"index\":[1,2,3],\"data\":[\"2022-08-19T12:22:27.000Z\",\"2022-08-19T12:22:42.000Z\",\"2022-08-19T12:22:57.000Z\"]}", "history_dates": "{\"name\":\"ds\",\"index\":[0,1,2,3,4],\"data\":[\"2022-08-19T12:22:12.000Z\",\"2022-08-19T12:22:27.000Z\",\"2022-08-19T12:22:42.000Z\",\"2022-08-19T12:22:57.000Z\",\"2022-08-19T12:23:12.000Z\"]}", "train_holiday_names": null, "start": 1660911732.0, "t_scale": 60.0, "holidays": null, "history": "{\"schema\":{\"fields\":[{\"name\":\"Time\",\"type\":\"datetime\"},{\"name\":\"ds\",\"type\":\"datetime\"},{\"name\":\"y\",\"type\":\"number\"},{\"name\":\"floor\",\"type\":\"integer\"},{\"name\":\"t\",\"type\":\"number\"},{\"name\":\"y_scaled\",\"type\":\"number\"}],\"pandas_version\":\"0.20.0\"},\"data\":[{\"Time\":\"2022-08-19T12:22:12.000Z\",\"ds\":\"2022-08-19T12:22:12.000Z\",\"y\":12984.220703125,\"floor\":0,\"t\":0.0,\"y_scaled\":0.9961340907},{\"Time\":\"2022-08-19T12:22:27.000Z\",\"ds\":\"2022-08-19T12:22:27.000Z\",\"y\":13007.275390625,\"floor\":0,\"t\":0.25,\"y_scaled\":0.9979028191},{\"Time\":\"2022-08-19T12:22:42.000Z\",\"ds\":\"2022-08-19T12:22:42.000Z\",\"y\":13007.275390625,\"floor\":0,\"t\":0.5,\"y_scaled\":0.9979028191},{\"Time\":\"2022-08-19T12:22:57.000Z\",\"ds\":\"2022-08-19T12:22:57.000Z\",\"y\":13034.611328125,\"floor\":0,\"t\":0.75,\"y_scaled\":1.0},{\"Time\":\"2022-08-19T12:23:12.000Z\",\"ds\":\"2022-08-19T12:23:12.000Z\",\"y\":13034.611328125,\"floor\":0,\"t\":1.0,\"y_scaled\":1.0}]}", "train_component_cols": "{\"schema\":{\"fields\":[{\"name\":\"additive_terms\",\"type\":\"integer\"},{\"name\":\"multiplicative_terms\",\"type\":\"integer\"}],\"pandas_version\":\"0.20.0\"},\"data\":[{\"additive_terms\":0,\"multiplicative_terms\":0}]}", "changepoints_t": [0.25, 0.5, 0.75], "seasonalities": [[], {}], "extra_regressors": [[], {}], "fit_kwargs": {"init": {"k": -6.071532473033588e-18, "m": 0.995977990713187, "sigma_obs": 5.179248179203414e-19, "delta": [0.008080120085662196, -0.008080120085661615, 0.008007917061588635], "beta": [0.0012139205110011893]}}, "params": {"k": [[0.00707491368008995]], "m": [[0.9961340907119132]], "delta": [[-0.007074913680090278, 0.008388723472258287, -0.008388723472257987]], "sigma_obs": [[5.080834326335655e-17]], "beta": [[9.250106198304635e-17]], "trend": [[0.9961340907119132, 0.9979028191319356, 0.9979028191319356, 1.0, 1.0]]}, "__prophet_version": "1.0"}
+{"growth": "linear", "n_changepoints": 3, "specified_changepoints": false, "changepoint_range": 0.8, "yearly_seasonality": "auto", "weekly_seasonality": "auto", "daily_seasonality": "auto", "seasonality_mode": "multiplicative", "seasonality_prior_scale": 10.0, "changepoint_prior_scale": 0.05, "holidays_prior_scale": 10.0, "mcmc_samples": 0, "interval_width": 0.8, "uncertainty_samples": 1000, "y_scale": 160529.142578125, "logistic_floor": false, "country_holidays": null, "component_modes": {"additive": ["additive_terms", "extra_regressors_additive"], "multiplicative": ["multiplicative_terms", "extra_regressors_multiplicative", "holidays"]}, "changepoints": "{\"name\":\"ds\",\"index\":[1,2,3],\"data\":[\"2022-09-05T06:29:27.000Z\",\"2022-09-05T06:29:42.000Z\",\"2022-09-05T06:29:57.000Z\"]}", "history_dates": "{\"name\":\"ds\",\"index\":[0,1,2,3,4],\"data\":[\"2022-09-05T06:29:12.000Z\",\"2022-09-05T06:29:27.000Z\",\"2022-09-05T06:29:42.000Z\",\"2022-09-05T06:29:57.000Z\",\"2022-09-05T06:30:12.000Z\"]}", "train_holiday_names": null, "start": 1662359352.0, "t_scale": 60.0, "holidays": null, "history": "{\"schema\":{\"fields\":[{\"name\":\"Time\",\"type\":\"datetime\"},{\"name\":\"ds\",\"type\":\"datetime\"},{\"name\":\"y\",\"type\":\"number\"},{\"name\":\"floor\",\"type\":\"integer\"},{\"name\":\"t\",\"type\":\"number\"},{\"name\":\"y_scaled\",\"type\":\"number\"}],\"pandas_version\":\"0.20.0\"},\"data\":[{\"Time\":\"2022-09-05T06:29:12.000Z\",\"ds\":\"2022-09-05T06:29:12.000Z\",\"y\":160478.576171875,\"floor\":0,\"t\":0.0,\"y_scaled\":0.9996850017},{\"Time\":\"2022-09-05T06:29:27.000Z\",\"ds\":\"2022-09-05T06:29:27.000Z\",\"y\":160478.576171875,\"floor\":0,\"t\":0.25,\"y_scaled\":0.9996850017},{\"Time\":\"2022-09-05T06:29:42.000Z\",\"ds\":\"2022-09-05T06:29:42.000Z\",\"y\":160501.712890625,\"floor\":0,\"t\":0.5,\"y_scaled\":0.9998291295},{\"Time\":\"2022-09-05T06:29:57.000Z\",\"ds\":\"2022-09-05T06:29:57.000Z\",\"y\":160501.712890625,\"floor\":0,\"t\":0.75,\"y_scaled\":0.9998291295},{\"Time\":\"2022-09-05T06:30:12.000Z\",\"ds\":\"2022-09-05T06:30:12.000Z\",\"y\":160529.142578125,\"floor\":0,\"t\":1.0,\"y_scaled\":1.0}]}", "train_component_cols": "{\"schema\":{\"fields\":[{\"name\":\"additive_terms\",\"type\":\"integer\"},{\"name\":\"multiplicative_terms\",\"type\":\"integer\"}],\"pandas_version\":\"0.20.0\"},\"data\":[{\"additive_terms\":0,\"multiplicative_terms\":0}]}", "changepoints_t": [0.25, 0.5, 0.75], "seasonalities": [[], {}], "extra_regressors": [[], {}], "fit_kwargs": {"init": {"k": 0.0005733287676392457, "m": 0.999697888605434, "sigma_obs": 5.62149955900674e-17, "delta": [-0.0005733287676388255, 0.0006351168106242873, -0.0006351168106245462], "beta": [-1.0906131418044258e-05]}}, "params": {"k": [[-6.651580440697051e-17]], "m": [[0.9996850017047505]], "delta": [[0.0005765113643145946, -0.0005765113643144664, 0.0006834818166837375]], "sigma_obs": [[5.344210819219706e-19]], "beta": [[-0.005584188130277588]], "trend": [[0.9996850017047505, 0.9996850017047505, 0.999829129545829, 0.999829129545829, 1.0]]}, "__prophet_version": "1.0"}
